+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+### Version 4.26.0 (3rd February 2021)
+#### Added
+- Added wrapper method `updateConversionValue:` method to `Adjust` API to allow updating SKAdNetwork conversion value via SDK API.
+
+#### Fixed
+- Fixed nullability warnings.
+
+---
+
 ### Version 4.25.2 (1st February 2021)
 #### Added
 - Added Facebook audience network ad revenue source string.